Refactor Address into multiple components
