+++ src/components/Forms/Tick.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 import styled from '@emotion/styled'
 
 const activeColourSwitch = props => (props.active ? '#5284FF' : '#B0BECF')
